+++ src/css/components/controls.css
@@ -1,0 +1,52 @@
+.zoom-controls {
+    position: absolute;
+    top: 2rem;
+    right: 2.5rem;
+    display: flex;
+    gap: 12px;
+    z-index: 100;
+    background: rgba(255, 255, 255, 0.9);
+    padding: 8px;
+    border-radius: 12px;
+    backdrop-filter: blur(4px);
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
+}
+
+.zoom-button, .theme-toggle {
+    background: white;
+    border: 1px solid #4f46e5;
+    color: #4f46e5;
+    width: 36px;
+    height: 36px;
+    border-radius: 8px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.zoom-button:hover, .theme-toggle:hover {
+    background: #4f46e5;
+    color: white;
+    transform: translateY(-2px);
+}
+
+/* Dark mode control styles */
+[data-theme="dark"] .zoom-controls {
+    background: rgba(30, 41, 59, 0.9);
+}
+
+[data-theme="dark"] .theme-toggle,
+[data-theme="dark"] .zoom-button {
+    background: #1e293b;
+    border-color: #818cf8;
+    color: #818cf8;
+}
+
+[data-theme="dark"] .theme-toggle:hover,
+[data-theme="dark"] .zoom-button:hover {
+    background: #818cf8;
+    color: #0f172a;
+}
