+++ src/css/components/modal.css
@@ -1,0 +1,146 @@
+.modal {
+    display: none;
+    position: fixed;
+    z-index: 1000;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.2);
+    overflow-y: auto;
+    backdrop-filter: blur(3px);
+}
+
+.modal-content {
+    background: var(--modal-bg);
+    margin: 2% auto;
+    padding: 0;
+    border: none;
+    width: 85%;
+    max-width: 800px;
+    border-radius: 16px;
+    box-shadow: 0 4px 20px var(--modal-shadow);
+    transform: translateY(0);
+    opacity: 1;
+    transition: all 0.2s ease;
+    animation: modalOpen 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+    max-height: 90vh;
+    overflow-y: auto;
+}
+
+.modal-content.closing {
+    transform: translateY(10px);
+    opacity: 0;
+}
+
+@keyframes modalSlideIn {
+    to {
+        transform: translateY(0);
+        opacity: 1;
+    }
+}
+
+@keyframes modalOpen {
+    from {
+        opacity: 0;
+        transform: scale(0.95) translateY(20px);
+    }
+    to {
+        opacity: 1;
+        transform: scale(1) translateY(0);
+    }
+}
+.modal-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 1.5rem;
+    padding: 1.5rem 2rem;
+    border-bottom: 1px solid #e9ecef;
+    background: var(--modal-header-bg);
+    margin: 0;
+    border-bottom: none;
+    border-radius: 16px 16px 0 0;
+    position: relative;
+    padding-right: 60px;
+    border-color: var(--border-color);
+}
+
+.modal-header h2 {
+    color: white;
+    font-size: 1.5rem;
+    font-weight: 600;
+}
+
+.close {
+    color: white;
+    position: absolute;
+    right: 20px;
+    top: 20px;
+    font-size: 28px;
+    cursor: pointer;
+    opacity: 0.9;
+    transition: all 0.2s ease;
+    text-shadow: 0 1px 2px rgba(0, 0, 0, 0.2);
+    z-index: 3;
+}
+
+.close:hover {
+    opacity: 1;
+    transform: scale(1.1);
+}
+
+.modal-body {
+    padding: 2rem;
+}
+
+#copyButton {
+    padding: 0.6rem 1.2rem;
+    background: white;
+    color: #4f46e5;
+    border: 2px solid #4f46e5;
+    border-radius: 8px;
+    font-size: 0.9rem;
+    font-weight: 500;
+    transition: all 0.2s ease;
+    z-index: 2;
+    position: relative;
+}
+
+#copyButton:hover {
+    background: #4f46e5;
+    border-color: #4f46e5;
+    color: white;
+    transform: translateY(-2px);
+    box-shadow: 0 2px 8px rgba(79, 70, 229, 0.3);
+}
+
+.details {
+    background: var(--modal-details-bg);
+    padding: 1.25rem;
+    border-radius: 12px;
+    border: 1px solid var(--modal-details-border);
+    margin-bottom: 1.5rem;
+}
+
+.details p {
+    margin: 0.75rem 0;
+    color: var(--modal-details-text);
+    font-size: 0.95rem;
+}
+
+.details strong {
+    color: #6366f1; 
+    min-width: 100px;
+    display: inline-block;
+    font-weight: 500;
+}
+
+.description {
+    color: #495057;
+    line-height: 1.6;
+    font-size: 0.95rem;
+    padding: 0.5rem 0 0.5rem 1.5rem;
+    border-left: 3px solid #6366f1;
+    margin-left: 0.5rem;
+}
