+++ src/index.html
@@ -1,34 +1,52 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
   <head>
+    <meta charset="UTF-8" />
+    <meta
+      name="description"
+      content="Interactive map of India showcasing states, capitals, and cultural information"
+    />
     <!-- External resources and meta tags -->
     <script src="https://d3js.org/d3.v7.min.js"></script>
-    <link rel="stylesheet" href="styles.css" />
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&display=swap" rel="stylesheet">
+
+    <link rel="stylesheet" href="css/main.css" />
+    <link rel="stylesheet" href="css/components/controls.css" />
+    <link rel="stylesheet" href="css/components/header.css" />
+    <link rel="stylesheet" href="css/components/map.css" />
+    <link rel="stylesheet" href="css/components/modal.css" />
+    <link rel="stylesheet" href="css/components/themes.css" />
+
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&display=swap"
+      rel="stylesheet"
+    />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css" />
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css"
+    />
     <title>Explore India - Interactive State Map</title>
   </head>
   <body>
     <!-- Header section -->
     <div class="header">
-        <h1>Explore India</h1>
-        <p>Click on any state to discover its rich culture and heritage</p>
+      <h1>Explore India</h1>
+      <p>Click on any state to discover its rich culture and heritage</p>
     </div>
 
     <!-- Map container with controls -->
     <div class="map-container">
-        <!-- SVG will be inserted here by JavaScript -->
-        <div class="zoom-controls">
-            <button class="theme-toggle">
-                <i class="fas fa-moon"></i>
-            </button>
-            <button class="zoom-button" onclick="window.indiaMap.resetZoom()">
-                <i class="fas fa-home"></i>
-            </button>
-        </div>
+      <!-- SVG will be inserted here by JavaScript -->
+      <div class="zoom-controls" role="group" aria-label="Map controls">
+        <button class="theme-toggle" aria-label="Toggle dark mode">
+          <i class="fas fa-moon" aria-hidden="true"></i>
+        </button>
+        <button class="zoom-button" onclick="window.indiaMap.resetZoom()">
+          <i class="fas fa-home"></i>
+        </button>
+      </div>
     </div>
 
     <!-- State information modal -->
@@ -37,7 +55,9 @@
         <span class="close">&times;</span>
         <div class="modal-header">
           <h2 id="stateName" style="margin: 0"></h2>
-          <button id="copyButton"><i class="fas fa-copy"></i> Copy Details</button>
+          <button id="copyButton">
+            <i class="fas fa-copy"></i> Copy Details
+          </button>
         </div>
         <div class="modal-body">
           <div class="details">
@@ -49,10 +69,10 @@
               <strong>Literacy Rate:</strong> <span id="stateLiteracy"></span>
             </p>
           </div>
-              <div class="description" id="stateDescription"></div>
+          <div class="description" id="stateDescription"></div>
         </div>
       </div>
     </div>
-    <script src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
   </body>
 </html>
