+++ src/js/config/mapConfig.js
@@ -1,0 +1,7 @@
+export const MAP_CONFIG = {
+    center: [78.9629, 22.5937],
+    scale: 900,
+    width: window.innerWidth - 60,
+    height: window.innerHeight - 180,
+    initialTransform: null
+};
